[REFACTOR] rename relationship file names to account contact
